requirements.txt
pika~=1.3.0
Flask~=2.2.2
psycopg2-binary
requests~=2.28.1
esanpy~=1.1.0
wordcloud~=1.8.2.2
elasticsearch~=8.4.3
pip
numpy~=1.23.4
matplotlib~=3.6.1
nltk~=3.7